social_network/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
djangorestframework==3.15.1
Faker==25.6.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.1
